\name{adu340_4small}
\alias{adu340_4small}
\docType{data}
\title{Calcium-imaging dataset using Fura-2}
\description{
  \code{adu340_4small} results from a calcium-buffering experiment performed in a glomerulus of the Cockroach (\emph{Periplaneta americana}) antennal lobe, where the projection neurons were imaged using Fura-2. Calcium variations were measured through time and space at 340 nm.}
\usage{data(adu340_4small)}
\format{
  The dataset is stored as an \code{array} of \code{integer}. \code{adu340_4small} contains three dimensions width, height and time.
}
\details{ \code{adu340_4small} contains a dataset recorded in vitro from a glomerulus of the Cockroach antennal lobe.}
\source{Debora Fusca (Kloppenburg Lab, University of Cologne, Germany) performed the experiment.}
\examples{
data(adu340_4small)
}
\keyword{datasets}
