\name{DynClust-package}
\alias{DynClust-package}
\docType{package}
\title{non-parametric denoising and clustering method of noisy images both indexed by time and space}
\description{The aim of the method is first to denoise each temporal sequence using both the spatial and temporal information contained in the data, and then to cluster them according to their dynamic features. The method is based on the assumption that a finite sequence of noisy images both indexed by time and space is composed of noisy versions of only a limited amount of dynamic features.}
\details{
\tabular{ll}{Package: \tab DynClust\cr
Type: \tab Package\cr
Title: \tab non-parametric denoising and clustering method of noisy images both indexed by time and space\cr
Version: \tab 1.0\cr
Date: \tab 2012-07-31\cr
Author: \tab Yves Rozenholc, Christophe Pouzat, Tiffany Lieury\cr
Maintainer: \tab Tiffany Lieury <lieuryt-dynclust@yahoo.fr>\cr
Depends: \tab R (>= 2.10)\cr
License: \tab GPL (>=2)\cr}
}
\author{Yves Rozenholc, Christophe Pouzat, Tiffany Lieury}

\keyword{ package }


