\name{getClusterMap}
\alias{getClusterMap}
\title{getClusterMap}
\description{'getClusteredMap' returns the location of each cluster as matrix}
\usage{getClusterMap(fp.dim, fp.res.clust)}
\arguments{
  \item{fp.dim}{dimension of the dataset}
  \item{fp.res.clust}{a list containing the results of the clustering procedure (cf. 'ClusteringFct')}
}



\author{Tiffany Lieury,Christophe Pouzat, Yves Rozenholc}





