\name{getDenoisedSet}
\alias{getDenoisedSet}
\title{getDenoisedSet}
\description{'getDenoisedSet' returns the denoised dataset as an array}
\usage{getDenoisedSet(fp.data.array, fp.res.listdenois)}
\arguments{
  \item{fp.data.array}{an 3D array corresponding to the dataset (the third dimension is the time)}
  \item{fp.res.listdenois}{a list containing the results of the denoising procedure (cf. 'callDenoiseVoxel')}
}

\value{an array containing the denoised version of the dataset}

\author{Tiffany Lieury,Christophe Pouzat, Yves Rozenholc}





