\name{updateList}
\alias{updateList}
\title{updateList}
\description{'updateList' eliminates the clustered pixels x time from 'pixtoclust'
a vector containing the indexes of the pixels left for clustering   }
\usage{updateList(fp.pixtoclust, fp.neighinclust)}
\arguments{
  \item{fp.pixtoclust}{a vector containing the indexes of the pixels left for clustering   }
  \item{fp.neighinclust}{a vector containing the indexes of the neighboors of the current pixel to clust}
}

\value{a vector containing the indexes of the pixels left for clustering}

\author{Yves Rozenholc, Christophe Pouzat, Tiffany Lieury}






\keyword{internal}

