\docType{data}
\name{adu340_4small}
\alias{adu340_4small}
\title{Calcium-imaging dataset using Fura-2}
\format{50x50x128 array of integers}
\usage{
data(adu340_4small)
}
\description{
adu340_4small is the result from a calcium-buffering
experiment performed in a glomerulus of the Cockroach
(\emph{Periplaneta americana}) antennal lobe, where the
projection neurons were imaged using Fura-2. Calcium
variations were measured through time and space at 340 nm.
}
\author{
Debora Fusca at Kloppenburg Lab --University of Cologne,
Germany
}

